+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added A11y guide and examples for `terra-button`.
+
 ## 1.20.0 - (February 7, 2023)
 
 * Changed